Create EmbeddedPaymentMethodsViewTests
