+++ StripePaymentSheet/StripePaymentSheetTests/PaymentSheet/EmbeddedPaymentMethodsViewSnapshotTests.swift
@@ -732,63 +732,6 @@ class EmbeddedPaymentMethodsViewSnapshotTests: STPSnapshotTestCase {
         verify(embeddedView)
     }
 
-    // MARK: EmbeddedPaymentMethodsViewDelegate test
-    func testEmbeddedPaymentMethodsView_heightDidChange() {
-        let mockMandateProvider = MockMandateProvider { paymentMethodType in
-            if paymentMethodType == .stripe(.cashApp) {
-                let longText = String(repeating: "This is a long mandate text. ", count: 20)
-                return NSAttributedString(string: longText)
-            }
-            if paymentMethodType == .stripe(.payPal) {
-                let mediumText = String(repeating: "This is a long mandate text. ", count: 10)
-                return NSAttributedString(string: mediumText)
-            }
-            if paymentMethodType == .stripe(.amazonPay) {
-                let shortText = String(repeating: "This is a long mandate text. ", count: 3)
-                return NSAttributedString(string: shortText)
-            }
-            return nil
-        }
-
-        let mockDelegate = MockEmbeddedPaymentMethodsViewDelegate()
-
-        let embeddedView = EmbeddedPaymentMethodsView(
-            initialSelection: nil,
-            paymentMethodTypes: [.stripe(.card), .stripe(.cashApp), .stripe(.klarna), .stripe(.payPal)],
-            savedPaymentMethod: nil,
-            appearance: .default,
-            shouldShowApplePay: true,
-            shouldShowLink: true,
-            savedPaymentMethodAccessoryType: .none,
-            mandateProvider: mockMandateProvider
-        )
-        embeddedView.delegate = mockDelegate
-        embeddedView.autosizeHeight(width: 300)
-
-        let rowButtons = embeddedView.stackView.arrangedSubviews.compactMap { $0 as? RowButton }
-        // Height did change should not be called after init
-        XCTAssertFalse(mockDelegate.didCallHeightDidChange)
-
-        // Simulate tapping Cash App
-        embeddedView.didTap(selectedRowButton: rowButtons[1], selection: .new(paymentMethodType: .stripe(.cashApp)))
-        XCTAssertTrue(mockDelegate.didCallHeightDidChange)
-        mockDelegate.reset()
-
-        // Simulate tapping Klarna
-        embeddedView.didTap(selectedRowButton: rowButtons[2], selection: .new(paymentMethodType: .stripe(.klarna)))
-        XCTAssertTrue(mockDelegate.didCallHeightDidChange)
-        mockDelegate.reset()
-
-        // Simulate tapping PayPal
-        embeddedView.didTap(selectedRowButton: rowButtons[3], selection: .new(paymentMethodType: .stripe(.payPal)))
-        XCTAssertTrue(mockDelegate.didCallHeightDidChange)
-        mockDelegate.reset()
-
-        // Simulate tapping PayPal again
-        embeddedView.didTap(selectedRowButton: rowButtons[3], selection: .new(paymentMethodType: .stripe(.payPal)))
-        XCTAssertFalse(mockDelegate.didCallHeightDidChange)
-    }
-
     func verify(
         _ view: UIView,
         identifier: String? = nil,
@@ -800,7 +743,7 @@ class EmbeddedPaymentMethodsViewSnapshotTests: STPSnapshotTestCase {
     }
 }
 
-private class MockMandateProvider: MandateTextProvider {
+class MockMandateProvider: MandateTextProvider {
     private let mandateResolver: (PaymentSheet.PaymentMethodType?) -> (NSAttributedString?)
 
     init(mandateResolver: @escaping (PaymentSheet.PaymentMethodType?) -> (NSAttributedString?)) {
@@ -813,18 +756,5 @@ private class MockMandateProvider: MandateTextProvider {
 
     func mandate(for paymentMethodType: PaymentSheet.PaymentMethodType?, savedPaymentMethod: STPPaymentMethod?, bottomNoticeAttributedString: NSAttributedString?) -> NSAttributedString? {
         return mandateResolver(paymentMethodType)
-    }
-}
-
-private class MockEmbeddedPaymentMethodsViewDelegate: EmbeddedPaymentMethodsViewDelegate {
-
-    private(set) var didCallHeightDidChange = false
-
-    func heightDidChange() {
-        didCallHeightDidChange = true
-    }
-
-    func reset() {
-        didCallHeightDidChange = false
     }
 }
